+++ src/components/LoginForm.tsx
@@ -6,7 +6,7 @@ import TextField from '@mui/material/TextField';
 import Button from '@mui/material/Button';
 import { useTypedSelector, useTypedDispatch } from '../hooks';
 import { logIn, resetRequestStatus } from '../slices/authSlice';
-import { ErrorMessage } from './';
+import { Message } from './';
 import type { SubmitHandler } from 'react-hook-form';
 
 interface Inputs {
@@ -65,9 +65,9 @@ const LoginForm = () => {
         </Grid>
         {requestStatus === 'failed' && (
           <Grid item xs={12}>
-            <ErrorMessage variant="caption">
+            <Message variant="caption">
               Something went wrong. Check if your e-mail / password is correct.
-            </ErrorMessage>
+            </Message>
           </Grid>
         )}
       </Grid>
