+++ src/components/index.tsx
@@ -7,7 +7,7 @@ import BouncingButton from './BouncingButton';
 import LoginForm from './LoginForm';
 import SignupForm from './SignupForm';
 import FilterForm from './FilterForm';
-import ErrorMessage from './ErrorMessage';
+import Message from './Message';
 import Drawings from './Drawings';
 import FeedList from './FeedList';
 
@@ -20,7 +20,7 @@ export {
   LoginForm,
   SignupForm,
   FilterForm,
-  ErrorMessage,
+  Message,
   Logo,
   LogoutButton,
   BouncingButton,
